FreeBSD: Fix OpenDKIM verification.
This was broken, but technically harmless for Chatmail to Chatmail communication

It is related to our workaround in filtermail where we changed the custom source
address from 127.0.0.2 to to 127.0.0.1 to fix socket bind errors, but the relay
design was using this trick to fool OpenDKIM into thinking it should be validating
the incoming mails (because it's not coming from 127.0.0.1/::1) as the first step
for incoming mail is to filter it for PGP, not to bother with DKIM validation. This
means all incoming mail was needlessly being signed, unless you were in a FreeBSD
jail without VNET as 127.0.0.1 isn't real // gets remapped to another IP.

OpenDKIM supports another method to signal when it should be signing vs verifying,
so we use that now.